deploy: 리뷰 관련 수정 버전 배포

+++ src/main/java/com/gamsa/review/dto/AnswerSaveRequest.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class AnswerSaveRequest {
 
     private final Integer questionId;
-    private final int score;
+    private final double score;
 }
